requirements.txt
Django==4.2.7
supabase==2.0.2
python-dotenv==1.0.0
requests==2.31.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.7
africastalking==1.2.6
